1、修改关于我们界面  版权符号标识 2、解决服务器风速为double 导致类型转换异常 3、修改待审批巡查计划列表排序方式（最新创建数据显示再最前面）

+++ app/src/main/java/com/isoftstone/smartsite/http/aqi/DataQueryBean.java
@@ -21,13 +21,13 @@ public class DataQueryBean {
     private double lightIntensity;//	Double 	光照强度
     private double atmosphericPressure;//	Double	大气压力
     private String windDirection = "";  //风向
-    private int windSpeed;     //风速
+    private double windSpeed;     //风速
 
-    public int getWindSpeed() {
+    public double getWindSpeed() {
         return windSpeed;
     }
 
-    public void setWindSpeed(int windSpeed) {
+    public void setWindSpeed(double windSpeed) {
         this.windSpeed = windSpeed;
     }
